Renamed ferrari solver file to remove mix precision distinction, Added file for 1010 algorithm
