+++ src/quartics/ferrari_qs.py
